Integration de de la mise a jour aprés supression/ajout d'un projet.

+++ FrontEnd/assets/scripts/works.js
@@ -1,3 +1,14 @@
+//fonction pour changer la page d'acceuil si le token est dans le localStorage
+function indexChanges() {
+    if (token) {
+        loginBtn.textContent = "logout";
+        modalEditBtn.style.display = "flex";
+        editMode.style.display = "flex";
+    }
+}
+indexChanges();
+
+//fonction pour récupérer les projets et les intégrer dans la galerie
 async function worksGet() {
     const works = await fetch("http://localhost:5678/api/works").then(
         (response) => response.json()
@@ -13,9 +24,9 @@ async function worksGet() {
         worksElement.appendChild(titleElement);
     }
 }
-
 worksGet();
 
+//fonction pour récupérer les categories du formulaire de la modale
 async function modalCategoriesGet() {
     const categories = await fetch("http://localhost:5678/api/categories").then(
         (response) => response.json()
@@ -27,14 +38,10 @@ async function modalCategoriesGet() {
         modalCategories.appendChild(categoriesOptions);
     }
 }
-
 modalCategoriesGet();
 
+//fonction pour récupérer les projets et les intégrer dans la galerie de la modale
 async function worksModalGet() {
-    if (!token) {
-        console.error("Token non trouvé");
-        return;
-    }
     const works = await fetch("http://localhost:5678/api/works").then(
         (response) => response.json()
     );
@@ -49,32 +56,46 @@ async function worksModalGet() {
         worksElement.appendChild(imgElement);
         worksElement.appendChild(iconElement);
 
-        
+        //bouton pour supprimer un projet
         iconElement.addEventListener("click", async function () {
-            console.log(worksElement.id);
-            const response = await fetch(
-                "http://localhost:5678/api/works/" + worksElement.id,
-                {
-                    method: "DELETE",
-                    headers: {
-                        Authorization: `Bearer ${token}`,
-                    },
-                }
-            );
-            if (!response.ok) {
-                throw new Error(
-                    `Erreur : ${response.status} ${response.statusText}`
+            if (!token) {
+                console.error("Token non trouvé");
+                return;
+            }
+            try {
+                const response = await fetch(
+                    "http://localhost:5678/api/works/" + worksElement.id,
+                    {
+                        method: "DELETE",
+                        headers: {
+                            Authorization: `Bearer ${token}`,
+                        },
+                    }
                 );
-            } else {
-                worksElement.remove();
+                if (!response.ok) {
+                    throw new Error(
+                        `Erreur : ${response.status} ${response.statusText}`
+                    );
+                } else {
+                    while (sectionFiche.hasChildNodes()) {
+                        sectionFiche.removeChild(sectionFiche.firstChild);
+                    }
+                    while (modalGallery.hasChildNodes()) {
+                        modalGallery.removeChild(modalGallery.firstChild);
+                    }
+                    worksGet();
+                    worksModalGet();
+                }
+            } catch (error) {
+                console.error("Erreur lors de la suppresion du projet:", error);
             }
         });
     }
 }
-
 worksModalGet();
 
-async function addWork(e) {
+//fonction pour ajouter un projet via le formulaire de la modale
+async function workAdd(e) {
     e.preventDefault();
     if (!token) {
         console.error("Token non trouvé");
@@ -93,30 +114,38 @@ async function addWork(e) {
             throw new Error(
                 `Erreur : ${response.status} ${response.statusText}`
             );
-        }
-        const data = await response.json();
-        if (data) {
-            console.log(data);
+        } else {
+            while (sectionFiche.hasChildNodes()) {
+                sectionFiche.removeChild(sectionFiche.firstChild);
+            }
+            while (modalGallery.hasChildNodes()) {
+                modalGallery.removeChild(modalGallery.firstChild);
+            }
+            worksGet();
+            worksModalGet();
         }
     } catch (error) {
-        console.error("Erreur lors de l'ajout de l'image:", error);
+        console.error("Erreur lors de l'ajout du projet:", error);
     }
 }
 
 modalForm.addEventListener("submit", (e) => {
-    addWork(e);
+    workAdd(e);
 });
 
 modalEditBtn.addEventListener("click", () => {
     modal.style.display = "flex";
 });
+
 modalExitBtn.addEventListener("click", () => {
     modal.style.display = "none";
 });
+
 modalReturnBtn.addEventListener("click", () => {
     modalAdd.style.display = "none";
     modalRemove.style.display = "block";
 });
+
 modalAddBtn.addEventListener("click", () => {
     modalRemove.style.display = "none";
     modalAdd.style.display = "block";
